# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  name: string;
  image: string;
  description: string;
  amazon_link: string;
}

export interface Post {
  id: number;
  title: string;
  slug: string;
  category: string;
  accent_color: string;
  intro_text: string;
  cover_image: string;
  products: Product[];
  conclusion_text: string;
  created_at: string;
}

export interface PostInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  accent_color: string;
  /** @minLength 1 */
  intro_text: string;
  /** @minLength 1 */
  cover_image: string;
  /** @minItems 1 */
  products: Product[];
  /** @minLength 1 */
  conclusion_text: string;
}

export interface PostUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  slug?: string;
  /** @minLength 1 */
  category?: string;
  /** @minLength 1 */
  accent_color?: string;
  /** @minLength 1 */
  intro_text?: string;
  /** @minLength 1 */
  cover_image?: string;
  /** @minItems 1 */
  products?: Product[];
  /** @minLength 1 */
  conclusion_text?: string;
}

export interface Category {
  name: string;
  count: number;
  image: string;
}

export interface LoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface SessionUser {
  username: string;
  role: string;
}

export interface DashboardSummary {
  total_posts: number;
  categories: number;
  latest_post: Post | null;
}

export type ListPostsParams = {
category?: string;
search?: string;
featured?: boolean;
};