# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: posts
    description: Public home decor journal content
  - name: auth
    description: Admin session authentication
  - name: dashboard
    description: Protected publishing dashboard
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List decor journal posts
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a journal post
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Invalid post
        "401":
          description: Authentication required
  /posts/{id}:
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a journal post
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "401":
          description: Authentication required
        "404":
          description: Post not found
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a journal post
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Authentication required
        "404":
          description: Post not found
  /posts/slug/{slug}:
    get:
      operationId: getPostBySlug
      tags: [posts]
      summary: Get a post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Post not found
  /categories:
    get:
      operationId: listCategories
      tags: [posts]
      summary: List available post categories with counts
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in to the publishing dashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "401":
          description: Invalid credentials
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out of the publishing dashboard
      responses:
        "204":
          description: Logged out
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the current admin session
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "401":
          description: Not authenticated
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard publishing summary
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          description: Authentication required
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Product:
      type: object
      properties:
        name:
          type: string
        image:
          type: string
        description:
          type: string
        amazon_link:
          type: string
      required: [name, image, description, amazon_link]
    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        category:
          type: string
        accent_color:
          type: string
        intro_text:
          type: string
        cover_image:
          type: string
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        conclusion_text:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - title
        - slug
        - category
        - accent_color
        - intro_text
        - cover_image
        - products
        - conclusion_text
        - created_at
    PostInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        accent_color:
          type: string
          minLength: 1
        intro_text:
          type: string
          minLength: 1
        cover_image:
          type: string
          minLength: 1
        products:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Product"
        conclusion_text:
          type: string
          minLength: 1
      required:
        - title
        - slug
        - category
        - accent_color
        - intro_text
        - cover_image
        - products
        - conclusion_text
    PostUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        accent_color:
          type: string
          minLength: 1
        intro_text:
          type: string
          minLength: 1
        cover_image:
          type: string
          minLength: 1
        products:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Product"
        conclusion_text:
          type: string
          minLength: 1
    Category:
      type: object
      properties:
        name:
          type: string
        count:
          type: integer
        image:
          type: string
      required: [name, count, image]
    LoginInput:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required: [username, password]
    SessionUser:
      type: object
      properties:
        username:
          type: string
        role:
          type: string
      required: [username, role]
    DashboardSummary:
      type: object
      properties:
        total_posts:
          type: integer
        categories:
          type: integer
        latest_post:
          anyOf:
            - $ref: "#/components/schemas/Post"
            - type: "null"
      required: [total_posts, categories, latest_post]