# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Post } from './post';

export interface DashboardSummary {
  total_posts: number;
  categories: number;
  latest_post: Post | null;
}