# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface PostInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  accent_color: string;
  /** @minLength 1 */
  intro_text: string;
  /** @minLength 1 */
  cover_image: string;
  /** @minItems 1 */
  products: Product[];
  /** @minLength 1 */
  conclusion_text: string;
}