# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  name: string;
  image: string;
  description: string;
  amazon_link: string;
}